+++ ipldsch_types.go
@@ -128,7 +128,7 @@ type _HashMapRoot struct {
 
 // Int matches the IPLD Schema type "Int".  It has int kind.
 type Int = *_Int
-type _Int struct{ x int }
+type _Int struct{ x int64 }
 
 // Link matches the IPLD Schema type "Link".  It has link kind.
 type Link = *_Link
